requirements.txt
python-telegram-bot==22.0
selenium==4.32.0
webdriver-manager==4.0.2
Pillow==11.2.1
python-dotenv==1.1.0
httpx==0.28.1
